Let's try that again
After digging around a little more, I think it would be beneficial to
have a distinct Rails API endpoint service that talks to a
Typescript/React frontend app over the network. I'll have to come up
with a way to run both services on different ports for local development
and to have the API service running continually while NGINX serves files
from the client, but that's a problem for another day.