Separate api and scheduling into different dirs
Since scheduling turns out to be a harder problem, I'll frame the recipe
scheduling problem as a variation of the job-shop problem and use
Google's constraint optimization tools to solve it. This commit
separates different backend services into "api/" which will contain the
GraphQL/database schemas and resolvers, and "scheduler/" which will
contain Python programs which use Google's tools to do the optimization
heavy-lifting.

https://en.wikipedia.org/wiki/Job_shop_scheduling
https://developers.google.com/optimization/