Working on defining recipe objective functions
In addition to minimizing makespan, which is the total time it would
take to cook every recipe, we also want to optimize for each recipe
ending at the same time. We can make an optimization function where we
want to minimize the difference between the end time of the last step of
each recipe.