client: local Apollo client cache example
After logging in or signing up a user, we'll store some metadata like
their name, email, etc in the client cache so we can (1) ensure that
they're properly logged in, and (2) be able to have personalized
messages. In the future, signup/login/logout API calls will exist that
will return authentication/session tokens, so we can persist state
across refreshes.