api: start login/signup implementation
Did a little research into how to store passwords and it sounds like
salted hashes (yum!) are the way to go. Modified the datamodel to
include an encryptedPassword field on the user as well as making
firstName and lastName not required (the only required fields for
signing up should be email and password. All other fields should be
configured in a profile page). Removed createUser resolver, since we'll
use signup as the interface to create users. Implemented password
hashing in new login/signup resolvers, without cookie management. Added
tests for the new resolvers and updated tests to reflect schema change.