api: experimenting with express-session
I'd like to authenticate users through a JWT that users obtain at
login/signup time. Users will be able to send the JWT to the api for
every request to verify authenticity as well as potentially providing a
way of restricted authorization in the future (admin tokens vs general
tokens). There are different methods of storing JWTs in the client.
LocalStorage and SessionStorage are out since they are very vulnerable
to XSS attacks, so we'll attempt to store it in a cookie. Cookies are
still vulnerable to CSRF attacks, but we can mitigate it by including a
CSRF token in the cookie. This commit experiments modifying the
graphql-yoga server with express-session in order to send cookies.