persistent login via /me endpoint
Storing a cookie on the client is nice, since it's persisted across
tabs/refreshes and the express middleware can keep a nice little session
store to hold onto IDs.

The problem is that the client "forgets" about the user on every page
refresh, since, although it has the cookie, the actual userId is stored
in the session store on the API server. We can still use the cookie for
authentication, but in addition send the API a request for our user
metadata given whether we have a cookie or not. If we don't have a
cookie, we're not logged in. If we have the cookie, we're logged in and
we can populate Apollo's cache with the user's metadata.

bweis helped nudge me towards this solution.