client: Add CreateRecipeModal
Added another button on the RecipesPage that allows a user to create a
new Recipe record with no steps. A button on the page brings up a modal
similar to the UserSessionModal, but with a schema for the fields
required by the createRecipe resolver. The UI still doesn't update when
a new recipe is created.